requirements.txt
Django>=1.6.2
celery[redis]
django-compressor>=1.3
django-localflavor>=1.0
django-mptt>=0.6.0
djangorestframework>=2.3.12
django-celery-email>=1.0.4
redis>=2.4.9
ranking>=0.3.1
python-ldap>=2.3.13
south
dealer>=1.1.0
requests
